requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
albumentations>=1.1.0
streamlit>=1.0.0
tqdm>=4.62.0
ultralytics>=8.0.0
requests>=2.0.0
flask
pathlib
pandas
scipy
zipfile36
imgaug
flask-cors
torchaudio
